generate Socket modules
There's a lot of common ffi code that we want to reuse across versions.
Previously we did this through the use of Moo and OO techniques.
Unfortunately such dynamic method dispatching has performance
implications. Considering the point of switching to FFI::Platypus is to
vastly improve ZMQ::FFI's performance we need to employ another
strategy.

To strike the balance between reuse and performance, we now add a dzil
build hook and generate the Socket modules with the common code injected
into each. As part of this:

    - update gitignore to ignore generated modules
    - add dzil hook to run generation script before build
    - update README HACKING instructions
    - add gen_modules.pl authordeps
    - install module generation script dependencies in .travis.yml