optimize Perl code based on profiling data
Try to eliminate as many speed bumps as possible when sending/receiving
millions of messages in a tight loop (zeromq weather station example is
a good demonstration of this, and was used for profiling purposes)

Compromise between performance and readability by only
aggressively optimizing the hot spots. Mostly it's just the send/recv
subs that need to be super aggressively optimized.

Minimize assignments and variable creation in subs.

Avoid scalar_to_buffer in ZMQ3 zmq_send. scalar_to_buffer is sllooww
(all that pack/unpack). Instead declare zmq_send with string type for
buffer param in ZMQ3.  This allows us to pass the message directly and
avoid the copy.  (zmq_send doesn't take control of the buffer, so this
is safe).

Don't get $self by shifting off the stack. Either directly reference or
assign using @_

Use a lookup table to cache native sizes

We're being a little less aggressive optimizing the ZMQ2 backend since
it's 2 generations old, and less used.  But we can always optimize it
more if we have to.