explicitly assign data arg in ZMQ2 send
Suppresses possible pack pointer warning "Attempt to pack pointer to
temporary value."  This warnings occurs if the user directly nests a
subroutine call as the argument to send().

In reality the warning is a false positive (since we don't return the
pointer to the temporary buffer), but there's no easy way to suppress
this outside of an extra assignment.