don't cleanup instances cloned by fork
if context/socket was cloned from a parent process we should _not_ attempt to
do cleanup. cloned instances are not in a valid state and cleanup is no longer
safe. cleanup should only be done from the originating process. set a pid field
that we can use to easily check if an instance is in the originating process.