add alternative error handling semantics
There are times users may not want exceptional error handling behavior.
For instance ZMQ_DONTWAIT can return EAGAIN, which is valid, non-fatal,
and should be handled.

So that it's not necessary to wrap cases like above in eval/Try::Tiny
blocks and do icky string comparisons on the error message, disabling
exceptional error behavior is no supported. The attribute die_on_error
is toggles the behavior.

After every zmq operation users can now call $socket->has_error to check
if an error occured, and either $socket->last_errno to get the error
code or $socket->last_strerror to get the error message.

Code was reorganized as follows to accomplish the above:
    * Kill ErrorHandler and turn ErrorHelper into the single Role with
      all the error related methods and attributes
    * Contexts and Sockets now consume ErrorHelper
    * Contexts and Sockets now attach the zmq error functions
      (zmq_errno, zmq_strerror) as xsubs instead of doing lazy dynamic
      ffi calls
    * Attaching the error ffi also prevents the OSX errno clobbering bug
      (c9b722ec) and makes the gymnastics around loading a separate
      ErrorHandler instance unecessary
    * To minimize duplicate code we put calls to the error ffi
      functions in the ErrorHelper role. We (judiciously) use symbolic
      refs to dynamically call the ffi sub associated with the current
      backend. This dynamic lookup is slow, but as the error helpers are
      not hotspots we can get away this.

Calls to the new error helpers have been optimized where possible, but
we still lose a smidgen of performance.  That's just the trade off for a
more fully featured API.