switch back to EV in fd.t
make EV 4 a test requires dependency.

this is a workaround for GH #17. not requiring EV at all means people
with EV 3 will get explosions. this way at least we ensure they upgrade
to EV 4 when installing ZMQ::FFI and that way sidestep the EV 3 issue.

using EV/nonblocking semantics with ZMQ::FFI is likely to be a common
enough use case that requiring it in our nonblocking test (and requiring
EV 4) seems reasonable.