Revert "switch back to EV in fd.t"
This reverts commit e595d7d77c5828fcddd8c778331fc0e2b8f1036c.

The EV version 3 thing (GH #17) turned out to be a red herring. The
segfault really has to do with race conditions during global
destruction.. for some reason using EV 3 just happens to trigger the
perfect storm of fail during global destruction.

So switch back to using condvar in the test and remove the EV 4 version
kludge (which doesn't fix the underlying issue anyways).

The more general segfault issue is now tracked in GH #19.