add Makefile.PL check for unthreaded BSD perls
This should resolve #13.

Follow @plicease's recommendation and use FFI::Platypus find_symbol on
the current perl process to see if the pthread_self symbol exists.  If
we're on BSD and the symbol doesn't exist bail on installing.

This is preferrable over a $Config{usethreads} check as this will work
even for custom perls that were manually linked to pthread (but not
compiled with the usethreads flag).