check thread id in cleanup routines
create util helper 'current_tid' that returns either the thread id if
this is a perl with threads enabled, or -1

following the pattern for fork, create an attribute _tid that is set at
object instantiation with the tid of the object's originating thread

in the cleanup routines (socket close, context destroy), check that the
current tid matches the object's tid, and only then proceed with cleanup

in the case of zmq_msg_close:

    for the thread case we only cleanup the underlying zmq_msg_t once in the
    socket's originating thread

    for the fork case we cleanup the underlying zmq_msg_t in every process

This is because with fork the underlying zmq_msg_t is actually copied,
and needs to be free'd in every process. However with threads the same
zmq_msg_t is pointed to from every thread, and should only be free'd
once.