create a bare docker image for testing installs
there was an issue where ZMQ::FFI was failing during install because
FFI::Platypus was required during configure (for a threaded perl check
on some BSDs), but FFI::Platypus was not listed as a configure
requirement and so was not installed before configure.  the current
docker images already have FFI::Platypus installed, so even though a
test install was done, this error was not caught.

instead two docker images are created now, a 'vanilla' perl-zmq-base
image that only includes gcc, make, libzmq, cpanm, and a few other base
os dependencies for testing installs, and the more complete image that has
dzil and all libzmq versions for more extensive unit testing.  this
should hopefully catch future install issues.

ref: #42